Move UITransformUtils & UITypes.
